Update uniswap-sdk-core-rust version and add test cases for price tick conversions
The version of uniswap-sdk-core-rust dependency in Cargo.toml is updated from 0.4.0 to 0.5.0 which is now sourced from Github repo. Changes in the price_tick_conversions.rs, include modifications in the calculation of sqrt_ratio and the addition of various tests for price and tick conversions. Cargo.lock is updated accordingly to reflect these changes.