Add utility to compute maximum liquidity for given amounts
This commit adds a new utility to compute the maximum amount of liquidity received for given amounts of two tokens. It provides two variants of computation, one precise and another imprecise, to accommodate both precise and fast computational needs. This utility will help in calculating how much liquidity will be received for given amounts of tokens at certain price boundaries. Tests are also added to verify the functionality of this module.