Enable optional inclusion of `extensions` module
The `extensions` module has been made optional so that it can be included or excluded during the build using feature flags. This is reflected in the GitHub Actions workflow file where the `--all-features` flag is now used during build and testing. The version of the Uniswap V3 SDK is also upgraded to 0.9.0.