Update README
The main changes included in this commit are the version update for `uniswap-v3-sdk` in the `Cargo.toml` snippet and an elaboration to the linting instructions. Additionally, the term 'module' has been replaced with 'feature' in reference to "extensions", and a new reference link to `uniswap-v3-automation-sdk` has been added.