Update data conversions with newly added `types` module
This commit introduces a new `types` module which is used to convert between ethers types and alloy types. This provides a more streamlined conversion mechanism and reduces dependencies on individual methods for conversion. This implementation involves changes throughout various parts of the codebase including sqrt price math, bit math, pool, and others. It also includes the addition of ethers-core as a dependency in Cargo.toml.