Update data conversions with newly added `types` module (#18)
* Update data conversions with newly added `types` module

This commit introduces a new `types` module which is used to convert between ethers types and alloy types. This provides a more streamlined conversion mechanism and reduces dependencies on individual methods for conversion. This implementation involves changes throughout various parts of the codebase including sqrt price math, bit math, pool, and others. It also includes the addition of ethers-core as a dependency in Cargo.toml.

* Update uniswap-v3-sdk version to 0.13.1

The version of uniswap-v3-sdk has been upgraded from 0.13.0 to 0.13.1. This is reflected in both the Cargo.toml and Cargo.lock files.