Add `Route` entity and upgrade dependencies (#25)
* Add `Route` entity and upgrade dependencies

This commit adds a new entity called Route which represents a list of pools through which a swap can occur, with associated constructor and methods. Along with this, the code also includes test cases for the new entity. Additionally, the version of the package has also been bumped up along with various dependency upgrades including tokio and toml_edit.

* Fix linting