Update comments and expand README for clarity (#31)
* Update comments and expand README for clarity

This commit revises documentation and comments across several files to improve clarity and provide additional information for users. It also marks one flaky test to be ignored. In the extensions section of the README, the modules' functions and their usage have been expanded for better understanding.

* Update uniswap-v3-sdk to version 0.23.1

The uniswap-v3-sdk version has been bumped up from 0.23.0 to 0.23.1. This reflects in both the Cargo.toml and Cargo.lock files.