chore(doc): Update code formatting and documentation (#41)
* chore(doc): Update code formatting and documentation

Reformatted code to adhere to a maximum line width of 100 characters and better comment wrapping. Updated package version and certain dependency versions. Updated access modifiers for utility functions and entity functions to public. Improved code readability and maintainability without changing functionality.

* Update GitHub Actions cache key in Rust workflow

The cache key in the Rust GitHub Actions workflow has been updated. Previously, it was based on hashing the Cargo.lock file, but now this key is based on hashing the Cargo.toml file, as it offers more consistency.