chore: Update dependency versions and refactor `CurrencyTrait` to `Currency`
This commit updates the versions of various dependencies in Cargo.toml as well as some other minor changes. The most significant change is the refactoring of `CurrencyTrait` to `Currency` in several files to better align this SDK's API with the latest Uniswap SDK. It streamlines imports and aligns the codebase with the new version of Uniswap SDK Core.