doc: add example and documentation for pool creation with tick data (#110)
* doc: add example and documentation for pool creation with tick data

This commit introduces an example demonstrating how to create a pool with a tick map data provider and simulates a swap in the README. Additionally, it includes documentation for the `from_pool_key_with_tick_data_provider` method in the codebase, enhancing clarity for developers utilizing the pool extension. These additions aim to improve understanding and usability of the library's pool-related features.

* Update examples/from_pool_key_with_tick_data_provider.rs

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* fmt

* Simplify Pool type by removing redundant generic.

Eliminate the unnecessary generic type parameter `<i32>` for `EphemeralTickMapDataProvider` in several places, streamlining the code and documentation. This change helps to reduce complexity and potential errors without affecting functionality. Additionally, improve code comments to enhance clarity and maintain accuracy.

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>